+++ README.md
@@ -24,7 +24,7 @@ Our website offers free, no-nonsense rank predictions for various entrance exams
 
 ## Currently Supported Exams
 
-- MET 2025 (Manipal Entrance Test)
+- MET 2026 (Manipal Entrance Test)
 - MIT GPA Calculator
 
 ## Methodology
